Module tray_rust::bxdf::bsdf

Defines the BSDF which acts as a container for composing the various BRDFs and BTDFs that describe the surface's properties

Structs

BSDF
The BSDF contains the various BRDFs and BTDFs that describe the surface's properties at some point. It also transforms incident and outgoing light directions into shading space to make the BxDFs easier to implement. TODO: We really need the memory pool. Each time we get the bsdf from a material we need to allocate a decent amount of stuff since they each need their own tangent, bitangent and differential geometry reference.