Enum tray_rust::bxdf::BxDFType

#[repr(u32)]

pub enum BxDFType {
    Reflection,
    Transmission,
    Diffuse,
    Glossy,
    Specular,
}

Various types of BxDFs that can be selected to specify which types of surface functions should be evaluated

Variants
ReflectionTransmissionDiffuseGlossySpecular
Methods

impl BxDFType

pub fn all_types() -> EnumSet<BxDFType>

Get an EnumSet containing all flags for the different types of BxDFs: Diffuse, Glossy, Specular

pub fn all_brdf() -> EnumSet<BxDFType>

Get an EnumSet containing all flags for reflective BxDFs (eg. BRDFs)

pub fn all_btdf() -> EnumSet<BxDFType>

Get an EnumSet containing all flags for transmissive BxDFs (eg. BTDFs)

pub fn all() -> EnumSet<BxDFType>

Get an EnumSet containing all flags for all BxDFs. This would be all types of BRDFs and BTDFs

pub fn specular() -> EnumSet<BxDFType>

Get an EnumSet containing flags for all types of specular BxDFs

pub fn non_specular() -> EnumSet<BxDFType>

Get an EnumSet containing flags for all non-specular BxDFs