Struct tray_rust::bxdf::fresnel::Dielectric

pub struct Dielectric {
    pub eta_i: f32,
    pub eta_t: f32,
}

Computes the Fresnel term for dielectric materials

Fields
eta_i: f32

Refractive index of the material the light is coming from

eta_t: f32

Refractive index of the material the light is hitting/entering

Methods

impl Dielectric

pub fn new(eta_i: f32, eta_t: f32) -> Dielectric

Create a new Dielectric Fresnel term for the boundary between two objects. eta_i: refractive index of the material the light is coming from. eta_t: refractive index of the material the light is entering.