Module tray_rust::bxdf::merl

The MERL BRDF represents the surface's properties through data loaded from a MERL BRDF Database file. The BRDF itself just stores the data loaded from the BRDF file while actual loading is done by the MERL material when it's created.

Structs

Merl
BRDF that uses measured data to model the surface reflectance properties. The measured data is from "A Data-Driven Reflectance Model", by Wojciech Matusik, Hanspeter Pfister, Matt Brand and Leonard McMillan, in ACM Transactions on Graphics 22, 3(2003), 759-769