Struct tray_rust::bxdf::merl::Merl

pub struct Merl {
    // some fields omitted
}

BRDF that uses measured data to model the surface reflectance properties. The measured data is from "A Data-Driven Reflectance Model", by Wojciech Matusik, Hanspeter Pfister, Matt Brand and Leonard McMillan, in ACM Transactions on Graphics 22, 3(2003), 759-769

Methods

impl Merl

fn new(brdf: Vec<f32>, n_theta_h: usize, n_theta_d: usize, n_phi_d: usize) -> Merl

Create a MERL BRDF to use data loaded from a MERL BRDF data file