Module tray_rust::bxdf::microfacet

Module providing various microfacet distribution functions and trait that's implemented by all provided distributions

Reexports

pub use self::beckmann::Beckmann;

pub use self::ggx::GGX;

Modules

beckmann
This module provides a Beckmann microfacet distribution with a Smith shadowing-masking term

ggx
This module provides a GGX microfacet distribution with a Smith shadowing-masking term. The GGX microfacet distribution is also sometimes referred to as Trowbridge-Reitz.

Traits

MicrofacetDistribution
Trait implemented by all microfacet distributions