Struct tray_rust::bxdf::microfacet::beckmann::Beckmann

pub struct Beckmann {
    // some fields omitted
}

Beckmann microfacet distribution with Smith shadowing-masking. This is the microfacet model described by Walter et al.

Methods

impl Beckmann

fn new(w: f32) -> Beckmann

Create a new Beckmann distribution with the desired width