Struct tray_rust::exec::distrib::master::Master

pub struct Master { /* fields omitted */ }

The Master organizes the set of Worker processes and instructions them what parts of the scene to render. As workers report results the master collects them and saves out the PNG once all workers have reported the frame.

Methods

impl Master

pub fn start_workers(
workers: Vec<String>,
config: Config,
img_dim: (usize, usize)
) -> (Master, EventLoop<Master>)

Create a new master that will contact the worker nodes passed and send instructions on what parts of the scene to start rendering