Trait tray_rust::exec::Exec

pub trait Exec {
    fn render(&mut self, scene: &mut Scene, rt: &mut RenderTarget, config: &Config);
}

Trait implemented by different execution environments that provides a method to call and render the scene, given the rendering arguments

Required Methods

fn render(&mut self, scene: &mut Scene, rt: &mut RenderTarget, config: &Config)

Render the scene using this rendering backend, will render out all frames of the image and save them out as instructed by the command line arguments TODO: In order to have a cleaner seperation we should pass more parameters to render. E.g. the scene. Or maybe a callback to a function that gets the frame's render target and can save it out?