Module tray_rust::film::animated_color

Provides an animated color value, so you can have colors change over time

Structs

AnimatedColor
AnimatedColor is a list of colors associated with time points in the scene that will compute the color at the desired time by blending the two nearest ones

ColorKeyframe
ColorKeyframe is a color associated with a specific time