Struct tray_rust::film::camera::Camera

pub struct Camera {
    pub active_at: usize,
    // some fields omitted
}

Our camera for the ray tracer, has a transformation to position it in world space

Fields

active_at

The frame this camera becomes active on

Methods

impl Camera

fn new(cam_world: AnimatedTransform, fov: f32, dims: (usize, usize), shutter_size: f32, active_at: usize) -> Camera

Create the camera with some orientation in the world specified by cam_world and a perspective projection with fov. The render target dimensions dims are needed to construct the raster -> camera transform animation is used to move the camera ote that this is specified in camera space where the camera is at the origin looking down the -z axis

fn animated_fov(cam_world: AnimatedTransform, fovs: Vec<f32>, fov_knots: Vec<f32>, fov_spline_degree: usize, dims: (usize, usize), shutter_size: f32, active_at: usize) -> Camera

Create a camera with some orientation in the world specified by cam_world and an animated perspective projection with fov. The render target dimensions dims are needed to construct the raster -> camera transform animation is used to move the camera ote that this is specified in camera space where the camera is at the origin looking down the -z axis

fn update_frame(&mut self, start: f32, end: f32)

Update the camera's shutter open/close time for this new frame

fn shutter_time(&self) -> (f32, f32)

Get the time that the shutter opens and closes at

fn generate_ray(&self, px: &(f32, f32), time: f32) -> Ray

Generate a ray from the camera through the pixel px