Struct tray_rust::film::color::Colorf

pub struct Colorf {
    pub r: f32,
    pub g: f32,
    pub b: f32,
    pub a: f32,
}

Colorf is a floating point RGBA color type a is typically used to store the weight of a color eg. in the render target for multisampling we need to track the weight to normalize in the end and is always initialized to 0

Fields
r: f32g: f32b: f32a: f32
Methods

impl Colorf

pub fn new(r: f32, g: f32, b: f32) -> Colorf

Create an RGB color

pub fn with_alpha(r: f32, g: f32, b: f32, a: f32) -> Colorf

Create an RGB color

pub fn broadcast(r: f32) -> Colorf

Create an RGB color using the same value for all channels

pub fn black() -> Colorf

Create a black color

pub fn clamp(&self) -> Colorf

Clamp the color values between [0, 1]

pub fn luminance(&self) -> f32

Compute the luminance of the color

pub fn is_black(&self) -> bool

Check if the color is black

pub fn has_nans(&self) -> bool

Check if any of the color channels are NaN

pub fn has_infs(&self) -> bool

Check if any of the color channels are infinite

pub fn to_srgb(&self) -> Colorf

Convert the linear RGB color to sRGB

pub fn exp(&self) -> Colorf

Return the color with values { e^r, e^g, e^b }