Trait tray_rust::film::filter::Filter

pub trait Filter {
    fn weight(&self, x: f32, y: f32) -> f32;

    fn width(&self) -> f32;

    fn inv_width(&self) -> f32;

    fn height(&self) -> f32;

    fn inv_height(&self) -> f32;
}

Trait implemented by all reconstructon filters. Provides methods for getting the width/height and computing the weight at some point relative to the filter center.

Required methods

fn weight(&self, x: f32, y: f32) -> f32

Compute the weight of this filter at some point (x, y) relative to the center of the filter

fn width(&self) -> f32

Return the width of the filter

fn inv_width(&self) -> f32

Return the inverse width of the filter

fn height(&self) -> f32

Return the height of the filter

fn inv_height(&self) -> f32

Return the inverse height of the filter

Loading content...