Struct tray_rust::film::render_target::RenderTarget

pub struct RenderTarget {
    // some fields omitted
}

RenderTarget is a RGBF render target to write our image too while rendering

Methods

impl RenderTarget

fn new(image_dim: (usize, usize), lock_size: (usize, usize), filter: Box<Filter + Send + Sync>) -> RenderTarget

Create a render target with width * height pixels

fn write(&self, samples: &[ImageSample], region: &Region)

Write all the image samples to the render target

fn clear(&mut self)

Clear the render target to black

fn dimensions(&self) -> (usize, usize)

Get the dimensions of the render target

fn get_render(&self) -> Vec<u8>

Convert the floating point color buffer to 24bpp sRGB for output to an image

fn get_rendered_blocks(&self) -> ((usize, usize), Vec<(usize, usize)>, Vec<f32>)

Get the blocks that have had pixels written too them. Returns the size of each block, a list of block positions in pixels and then pixels for the blocks (in a single f32 vec). The block's pixels are stored in the same order their position appears in the block positions vec and contain dim.0 * dim.1 * 4 f32's per block.

fn get_renderf32(&self) -> Vec<f32>

Get the raw floating point framebuffer