Struct tray_rust::geometry::animated_mesh::AnimatedMesh

pub struct AnimatedMesh { /* fields omitted */ }

An animated mesh composed of a series of meshes linearly interpolated between over time. It's assumed the mesh topology does not change.

Methods

impl AnimatedMesh

pub fn new(meshes: Vec<Arc<Mesh>>, times: Vec<f32>) -> AnimatedMesh

Create a new AnimatedMesh from the meshes passed. It's assumed the meshes are sorted in ascending time