Struct tray_rust::geometry::animated_mesh::AnimatedTriangle

pub struct AnimatedTriangle { /* fields omitted */ }

An animated triangle in the mesh. Just stores a reference to the mesh and the indices of each vertex

Methods

impl AnimatedTriangle

pub fn new(
a: usize,
b: usize,
c: usize,
data: Arc<AnimatedMeshData>
) -> AnimatedTriangle

Create a new triangle representing a triangle within the mesh passed