Struct tray_rust::geometry::emitter::Emitter

pub struct Emitter {
    pub emission: AnimatedColor,
    pub tag: String,
    // some fields omitted
}

An instance of geometry in the scene that receives and emits light.

Fields

emission

The light intensity emitted

tag

Tag to identify the instance

Methods

impl Emitter

fn area(geom: Arc<SampleableGeom + Send + Sync>, material: Arc<Material + Send + Sync>, emission: AnimatedColor, transform: AnimatedTransform, tag: String) -> Emitter

Create a new area light using the geometry passed to emit light TODO: We need sample methods for geometry to do this We also need MIS in the path tracer's direct light sampling so we get good quality

fn point(transform: AnimatedTransform, emission: AnimatedColor, tag: String) -> Emitter

Create a point light at the origin that is transformed by transform to its location in the world

fn intersect(&self, ray: &mut Ray) -> Option<(DifferentialGeometry, &Material)>

Test the ray for intersection against this insance of geometry. returns Some(Intersection) if an intersection was found and None if not. If an intersection is found ray.max_t will be set accordingly

fn radiance(&self, w: &Vector, _: &Point, n: &Normal, time: f32) -> Colorf

Return the radiance emitted by the light in the direction w from point p on the light's surface with normal n

fn get_transform(&self) -> &AnimatedTransform

Get the transform to place the emitter into world space

fn set_transform(&mut self, transform: AnimatedTransform)

Set the transform to place the emitter into world space