Enum tray_rust::geometry::instance::Instance

pub enum Instance {
    Emitter(Emitter),
    Receiver(Receiver),
}

Defines an instance of some geometry with its own transform and material

Variants

Emitter(Emitter)

Receiver(Receiver)

Methods

impl Instance

pub fn receiver(
geom: Arc<dyn BoundableGeom + Send + Sync>,
material: Arc<dyn Material + Send + Sync>,
transform: AnimatedTransform,
tag: String
) -> Instance

Create an instance of the geometry in the scene that will only receive light.

pub fn area_light(
geom: Arc<dyn SampleableGeom + Send + Sync>,
material: Arc<dyn Material + Send + Sync>,
emission: AnimatedColor,
transform: AnimatedTransform,
tag: String
) -> Instance

Create an instance of the geometry in the scene that will emit and receive light

pub fn point_light(
transform: AnimatedTransform,
emission: AnimatedColor,
tag: String
) -> Instance

Create a point light at the origin that is transformed by transform to its location in the world

pub fn intersect(&self, ray: &mut Ray) -> Option<Intersection>

Test the ray for intersection against this insance of geometry. returns Some(Intersection) if an intersection was found and None if not. If an intersection is found ray.max_t will be set accordingly

pub fn tag(&self) -> &str

Get the tag for this instance

pub fn get_transform(&self) -> &AnimatedTransform

Get the transform for this instance

pub fn set_transform(&mut self, transform: AnimatedTransform)

Set the transform for this instance