Struct tray_rust::geometry::intersection::Intersection

pub struct Intersection<'a, 'b> {
    pub dg: DifferentialGeometry<'a>,
    pub instance: &'b Instance,
    pub material: &'b Material,
}

Stores information about an intersection that occured with some instance of geometry in the scene

Fields

dg

The differential geometry holding information about the piece of geometry that was hit

instance

The instance of geometry that was hit

material

The material of the instance that was hit

Methods

impl<'a, 'b> Intersection<'a, 'b>

fn new(dg: DifferentialGeometry<'a>, inst: &'b Instance, mat: &'b Material) -> Intersection<'a, 'b>

Construct the Intersection from a potential hit stored in a Option. Returns None if dg is None or if the instance member of dg is None