Struct tray_rust::geometry::mesh::Mesh

pub struct Mesh {
    pub bvh: BVH<Triangle>,
}

A mesh composed of triangles, specified by directly passing the position, normal and index buffers for the triangles making up the mesh

Fields
bvh: BVH<Triangle>
Methods

impl Mesh

pub fn new(
positions: Arc<Vec<Point>>,
normals: Arc<Vec<Normal>>,
texcoords: Arc<Vec<Point>>,
indices: Vec<u32>
) -> Mesh

Create a new Mesh from the triangles described in the buffers passed This data could come from an OBJ file via tobj for example.

pub fn load_obj(file_name: &Path) -> HashMap<String, Arc<Mesh>>

Load all the meshes defined in an OBJ file and return them in a hashmap that maps the model's name in the file to its loaded mesh. TODO: Don't build the BVH until we actually use the mesh in the scene, will reduce scene load time. TODO: Currently materials are ignored