Struct tray_rust::geometry::receiver::Receiver

pub struct Receiver {
    pub material: Arc<Material + Send + Sync>,
    pub tag: String,
    // some fields omitted
}

An instance of geometry in the scene that only receives light

Fields

material

The material being used by this instance.

tag

Tag to identify the instance

Methods

impl Receiver

fn new(geom: Arc<BoundableGeom + Send + Sync>, material: Arc<Material + Send + Sync>, transform: AnimatedTransform, tag: String) -> Receiver

Create a new instance of some geometry in the scene

fn intersect(&self, ray: &mut Ray) -> Option<(DifferentialGeometry, &Material)>

Test the ray for intersection against this insance of geometry. returns Some(Intersection) if an intersection was found and None if not. If an intersection is found ray.max_t will be set accordingly

fn get_transform(&self) -> &AnimatedTransform

Get the transform to place the receiver into world space

fn set_transform(&mut self, transform: AnimatedTransform)

Set the transform to place the receiver into world space