Trait tray_rust::geometry::Boundable

pub trait Boundable {
    fn bounds(&self, start: f32, end: f32) -> BBox;

    fn update_deformation(&mut self, start: f32, end: f32) { ... }
}

Trait implemented by scene objects that can report an AABB describing their bounds

Required methods

fn bounds(&self, start: f32, end: f32) -> BBox

Get an AABB reporting the object's bounds over the time period The default implementation assumes the object isn't animated and simply returns its bounds. This is kind of a hack to use the BVH for animated geomtry (instances) and non-animated geometry (triangles).

Loading content...
Provided methods

fn update_deformation(&mut self, start: f32, end: f32)

Have the object recompute its bounds for the time range. In the case of deforming geometry this can rebuild acceleration structures for example.

Loading content...