Struct tray_rust::integrator::path::Path

pub struct Path { /* fields omitted */ }

The path integrator implementing Path tracing with explicit light sampling

Methods

impl Path

pub fn new(min_depth: u32, max_depth: u32) -> Path

Create a new path integrator with the min and max length desired for paths