Struct tray_rust::integrator::whitted::Whitted

pub struct Whitted { /* fields omitted */ }

The Whitted integrator implementing the Whitted recursive ray tracing algorithm

Methods

impl Whitted

pub fn new(max_depth: u32) -> Whitted

Create a new Whitted integrator with the desired maximum recursion depth for rays