Struct tray_rust::linalg::animated_transform::AnimatedTransform

pub struct AnimatedTransform { /* fields omitted */ }

An animated transform that blends between the keyframes in its transformation list over time.

Methods

impl AnimatedTransform

fn with_keyframes(keyframes: Vec<Keyframe>, knots: Vec<f32>, degree: usize) -> AnimatedTransform

Create an animated transformation blending between the passed keyframes

fn unanimated(transform: &Transform) -> AnimatedTransform

fn transform(&self, time: f32) -> Transform

Compute the transformation matrix for the animation at some time point using B-Spline interpolation.

fn animation_bounds(&self, b: &BBox, start: f32, end: f32) -> BBox

Compute the bounds of the box moving through the animation sequence by sampling time

fn is_animated(&self) -> bool

Check if the transform is actually animated