Struct tray_rust::linalg::keyframe::Keyframe

pub struct Keyframe {
    pub translation: Vector,
    pub rotation: Quaternion,
    pub scaling: Vector,
}

A transformation associated with a specific point in time. Note that this transform is now more implicit since they keyframe's times are stored as knots in the b-spline animation path

Fields

translation

rotation

scaling

Methods

impl Keyframe

fn new(transform: &Transform) -> Keyframe

Construct a new keyframe transformation, The transform will be stored in a decomposed form, M = TRS.

fn from_parts(translation: &Vector, rotation: &Quaternion, scaling: &Vector) -> Keyframe

Construct the keyframe from the decomposed transformation

fn transform(&self) -> Transform

Return the transformation stored for this keyframe