Struct tray_rust::linalg::matrix4::Matrix4

pub struct Matrix4 {
    // some fields omitted
}

Matrix4 is a 4x4 matrix stored in row-major format

Methods

impl Matrix4

fn zero() -> Matrix4

Return the zero matrix

fn identity() -> Matrix4

Return the identity matrix

fn new(mat: [f32; 16]) -> Matrix4

Create the matrix using the values passed

fn at(&self, i: usize, j: usize) -> &f32

Access the element at row i column j

fn at_mut(&mut self, i: usize, j: usize) -> &mut f32

Mutably access the element at row i column j

fn transpose(&self) -> Matrix4

Compute and return the transpose of this matrix

fn inverse(&self) -> Matrix4

Compute and return the inverse of this matrix

fn iter(&self) -> Iter<f32>

Return an iterator over the matrix's elements. The iterator goes row by row through the matrix.

fn has_nans(&self) -> bool