Struct tray_rust::linalg::point::Point

pub struct Point {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Point is a standard 3 component point but transforms as a point point when transformations are applied

Fields

x

y

z

Methods

impl Point

fn new(x: f32, y: f32, z: f32) -> Point

Initialize the point and set values for x, y, z

fn broadcast(x: f32) -> Point

Initialize the point with the same value of x, y, z

fn origin() -> Point

Initialize a point to be all 0 values

fn distance_sqr(&self, a: &Point) -> f32

Compute the squared distance between this point and another

fn distance(&self, a: &Point) -> f32

Compute the distance between this point and another