Struct tray_rust::linalg::quaternion::Quaternion

pub struct Quaternion {
    pub v: Vector,
    pub w: f32,
}

Quaternions describe a rotation in 3d space but can be properly interpolated unlike rotation matrices. The quaternion also provides methods to compute the corresponding rotation transform so that it can be composed with scaling and translation

Fields
v: Vectorw: f32
Methods

impl Quaternion

pub fn identity() -> Quaternion

Construct an identity quaternion

pub fn from_matrix(m: &Matrix4) -> Quaternion

Construct a quaternion from the rotation matrix Based on Shoemake 1991

pub fn from_transform(t: &Transform) -> Quaternion

Construct the quaternion from the transform

pub fn to_matrix(&self) -> Matrix4

Get the rotation transform described by this quaternion

pub fn to_transform(&self) -> Transform

Get the rotation transform described by this quaternion

pub fn normalized(&self) -> Quaternion

Get the normalized quaternion for this rotation