Struct tray_rust::linalg::ray::Ray

pub struct Ray {
    pub o: Point,
    pub d: Vector,
    pub min_t: f32,
    pub max_t: f32,
    pub depth: u32,
    pub time: f32,
}

Ray is a standard 3D ray, starting at origin o and heading in direction d The min and max points along the ray can be specified with min_t and max_t depth is the recursion depth of the ray

Fields

o

Origin of the ray

d

Direction the ray is heading

min_t

Point along the ray that the actual ray starts at, p = o + min_t * d

max_t

Point along the ray at which it stops, will be inf if the ray is infinite

depth

Recursion depth of the ray

time

Time point sampled by this ray

Methods

impl Ray

fn new(o: &Point, d: &Vector, time: f32) -> Ray

Create a new ray from o heading in d with infinite length

fn segment(o: &Point, d: &Vector, min_t: f32, max_t: f32, time: f32) -> Ray

Create a new segment ray from o + min_t * d to o + max_t * d

fn child(&self, o: &Point, d: &Vector) -> Ray

Create a child ray from the parent starting at o and heading in d

fn child_segment(&self, o: &Point, d: &Vector, min_t: f32, max_t: f32) -> Ray

Create a child ray segment from o + min_t * d to o + max_t * d

fn at(&self, t: f32) -> Point

Evaulate the ray equation at some t value and return the point returns result of self.o + t * self.d