Struct tray_rust::linalg::vector::Vector

pub struct Vector {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Vector is a standard 3 component vector

Fields
x: f32 y: f32 z: f32
Methods

impl Vector

pub fn new(x: f32, y: f32, z: f32) -> Vector

Initialize the vector and set values for x, y, z

pub fn broadcast(x: f32) -> Vector

Initialize the vector with the same value of x, y, z

pub fn length_sqr(&self) -> f32

Compute the squared length of the vector

pub fn length(&self) -> f32

Compute the length of the vector

pub fn normalized(&self) -> Vector

Get a normalized copy of this vector