Module tray_rust::material::matte

Defines a matte material used to describe diffuse materials

Scene Usage Example

The matte material requires a diffuse color and a roughness for the material. A roughness of 0 will select a Lambertian model while a roughness > 0 will select an Oren-Nayar reflectance model.

"materials": [
    {
        "name": "purple_matte",
        "type": "matte",
        "diffuse": [1, 0, 1],
        "roughness" 0.5
    },
    ...
]

Structs

Matte
The Matte material describes diffuse materials with either a Lambertian or Oren-Nayar BRDF. The Lambertian BRDF is used for materials with no roughness while Oren-Nayar is used for those with some roughness.