Struct tray_rust::material::merl::Merl

pub struct Merl {
    // some fields omitted
}

Material that uses measured data to model the surface reflectance properties. The measured data is from "A Data-Driven Reflectance Model", by Wojciech Matusik, Hanspeter Pfister, Matt Brand and Leonard McMillan, in ACM Transactions on Graphics 22, 3(2003), 759-769

Methods

impl Merl

fn load_file(path: &Path) -> Merl

Create a new MERL BRDF by loading the refletance data from a MERL BRDF database file