Struct tray_rust::material::specular_metal::SpecularMetal

pub struct SpecularMetal {
    // some fields omitted
}

The Specular Metal material describes specularly reflective metals using their refractive index and absorption coefficient

Methods

impl SpecularMetal

fn new(eta: &Colorf, k: &Colorf) -> SpecularMetal

Create a new specular metal with the desired metal properties. eta: refractive index of the metal k: absorption coefficient of the metal