Function tray_rust::mc::concentric_sample_disk

pub fn concentric_sample_disk(u: &(f32, f32)) -> (f32, f32)

Compute concentric sample positions on a unit disk mapping input from range [0, 1) to sample positions on a disk samples should be two random samples in range [0, 1) See: Shirley and Chiu, A Low Distortion Map Between Disk and Square