Function tray_rust::mc::uniform_sample_cone_frame

pub fn uniform_sample_cone_frame(
    samples: &(f32, f32), 
    cos_theta_max: f32, 
    w_x: &Vector, 
    w_y: &Vector, 
    w_z: &Vector
) -> Vector

Uniformly sample a direction in a cone with max angle cos_theta_max where the cone looks down the w_z vector provided, with w_x, w_y forming the rest of the coordinate frame for the cone