Module tray_rust::partition

Provides a general partitioning function that implements C++'s std::partition

Functions

partition
Re-orders elements in the range yielded by it based on pred. All elements that the predicate returns true for will be placed before all elements that the predicate returned false for. Also returns the index of the first element in the false group