Module tray_rust::sampler::ld

Provides a high quality sampling scheme based on (0, 2)-sequences See sec. 7.4.3 of Physically Based Rendering

Structs

LowDiscrepancy
Low discrepancy sampler that makes use of the (0, 2) sequence to generate well distributed samples

Functions

sample_02
Generate a sample from a scrambled (0, 2) sequence

sample_1d
Generate a 1D pattern of low discrepancy samples to fill the slice sample values will be normalized between [0, 1]

sample_2d
Generate a 2D pattern of low discrepancy samples to fill the slice sample values will be normalized between [0, 1]

sobol
Generate a scrambled Sobol' sequence value as described by Kollig & Keller (2002) and in PBR method is specialized for base 2

van_der_corput
Generate a scrambled Van der Corput sequence value as described by Kollig & Keller (2002) and in PBR method is specialized for base 2