Struct tray_rust::sampler::uniform::Uniform

pub struct Uniform { /* fields omitted */ }

Uniform sampler that takes one sample per pixel at the center of each pixel

Methods

impl Uniform

pub fn new(dim: (u32, u32)) -> Uniform

Create a uniform sampler to sample the image in dim.0 * dim.1 sized blocks