Struct tray_rust::scene::Scene

pub struct Scene {
    pub cameras: Vec<Camera>,
    pub bvh: BVH<Instance>,
    pub integrator: Box<Integrator + Send + Sync>,
    // some fields omitted
}

The scene containing the objects and camera configuration we'd like to render, shared immutably among the ray tracing threads

Fields

cameras

bvh

integrator

Methods

impl Scene

fn load_file(file: &str) -> (Scene, RenderTarget, usize, FrameInfo)

fn intersect(&self, ray: &mut Ray) -> Option<Intersection>

Test the ray for intersections against the objects in the scene. Returns Some(Intersection) if an intersection was found and None if not.

fn update_frame(&mut self, frame: usize, start: f32, end: f32)

Advance the time the scene is currently displaying to the time range passed

fn active_camera(&self) -> &Camera

Get the active camera for the current frame