Struct tray_rust::exec::distrib::worker::Worker

pub struct Worker {
    pub render_target: RenderTarget,
    pub scene: Scene,
    pub config: Config,
    // some fields omitted
}

A worker process for distributed rendering. Accepts instructions from the master process telling it what to render, after each frame is finished results are sent back to the master and the next frame is started. Once all frames are finished the worker exits

Fields

render_target

Render target the worker will write the current frame too

scene

config

Methods

impl Worker

fn listen_for_master(num_threads: u32) -> Worker

Listen on the worker PORT for the master to contact us and send us instructions about the scene we should render and what parts of it we've been assigned

fn send_results(&mut self)

Send our blocks back to the master