Module tray_rust::film

The film module provides color types and a render target that the image is written too.

Reexports

pub use self::color::Colorf;

pub use self::render_target::RenderTarget;

pub use self::camera::Camera;

pub use self::render_target::ImageSample;

pub use self::animated_color::{ColorKeyframe, AnimatedColor};

pub use self::image::Image;

Modules

animated_color
Provides an animated color value, so you can have colors change over time

camera
Provides a camera based on a single transformation that positions it in the scene

color
Defines types for operating with floating point and 8 bit RGB colors

filter
The filter module provides reconstruction filters to be used when writing samples to the render target. The filter width and height refer to how many pixels the filter covers, where a single pixel is 0.5x0.5

image
Provides a simple RGBA_F32 image, used by the distributed master to store results from the worker processes

render_target
Defines the render target for tray, where our image will be written too during rendering

Structs

FrameInfo
Struct to store various parameters for the frame timing