Struct tray_rust::film::color::Colorf

pub struct Colorf {
    pub r: f32,
    pub g: f32,
    pub b: f32,
    pub a: f32,
}

Colorf is a floating point RGBA color type a is typically used to store the weight of a color eg. in the render target for multisampling we need to track the weight to normalize in the end and is always initialized to 0

Fields

r

g

b

a

Methods

impl Colorf

fn new(r: f32, g: f32, b: f32) -> Colorf

Create an RGB color

fn broadcast(r: f32) -> Colorf

Create an RGB color using the same value for all channels

fn black() -> Colorf

Create a black color

fn clamp(&self) -> Colorf

Clamp the color values between [0, 1]

fn luminance(&self) -> f32

Compute the luminance of the color

fn is_black(&self) -> bool

Check if the color is black

fn has_nans(&self) -> bool

Check if any of the color channels are NaN

fn has_infs(&self) -> bool

Check if any of the color channels are infinite

fn to_srgb(&self) -> Colorf

Convert the linear RGB color to sRGB

fn exp(&self) -> Colorf

Return the color with values { er, eg, eb }