Struct tray_rust::film::render_target::ImageSample

pub struct ImageSample {
    pub x: f32,
    pub y: f32,
    pub color: Colorf,
}

A struct containing results of an image sample where a ray was fired through continuous pixel coordinates [x, y] and color color was computed

Fields
x: f32 y: f32 color: Colorf
Methods

impl ImageSample

pub fn new(x: f32, y: f32, color: Colorf) -> ImageSample