Struct tray_rust::geometry::bbox::BBox

pub struct BBox {
    pub min: Point,
    pub max: Point,
}

A box between the min and max points

Fields

min

max

Methods

impl BBox

fn new() -> BBox

Create a new degenerate box

fn singular(p: Point) -> BBox

Create a new box containing only the point passed

fn span(min: Point, max: Point) -> BBox

Create a new box spanning [min, max]

fn box_union(&self, b: &BBox) -> BBox

Get a box representing the union of this box with the one passed

fn point_union(&self, p: &Point) -> BBox

Get a box that contains the passed point, by expanding this box to reach the point

fn max_extent(&self) -> Axis

Compute the axis along which the box is longest

fn lerp(&self, tx: f32, ty: f32, tz: f32) -> Point

Compute the point in the box at some t value along each axis

fn offset(&self, p: &Point) -> Vector

Find the position of the point relative to the box, with min being the origin

fn surface_area(&self) -> f32

Compute the surface area of the box

fn fast_intersect(&self, r: &Ray, inv_dir: &Vector, neg_dir: &[usize; 3]) -> bool

Optimized ray-box intersection test, for use in the BVH traversal where we have pre-computed the ray's inverse direction and which directions are negative, indicated by a 1 for negative and 0 for non-negative Returns true if the box was hit