Struct tray_rust::geometry::bvh::BVH

pub struct BVH<T: Boundable> {
    // some fields omitted
}

A standard BVH2 that stores objects that can report their bounds in some space via the Boundable trait. The BVH is constructed using a SAH partitioning scheme

Methods

impl<T: Boundable> BVH<T>

fn unanimated(max_geom: usize, geometry: Vec<T>) -> BVH<T>

Create a new non-animated BVH holding the geometry

fn new(max_geom: usize, geometry: Vec<T>, start: f32, end: f32) -> BVH<T>

Create a new BVH using a SAH construction algorithm which holds the scene geometry for some time period.

fn rebuild(&mut self, start: f32, end: f32)

Re-build the BVH for the time range passed

fn intersect<'a, F, R>(&'a self, ray: &mut Ray, f: F) -> Option<R> where F: Fn(&mut Ray, &'a T) -> Option<R>

Traverse the BVH and call the function passed on the objects in the leaf nodes of the BVH, returning the value returned by the function after traversal completes

fn iter(&self) -> Iter<T>