Struct tray_rust::geometry::differential_geometry::DifferentialGeometry

pub struct DifferentialGeometry<'a> {
    pub p: Point,
    pub n: Normal,
    pub ng: Normal,
    pub dp_du: Vector,
    pub dp_dv: Vector,
    pub geom: &'a Geometry + 'a,
}

Stores information about a hit piece of geometry of some object in the scene

Fields
p: Point

The hit point

n: Normal

The shading normal

ng: Normal

The geometry normal

dp_du: Vector

Derivative of the point with respect to the u parameterization coord of the surface

dp_dv: Vector

Derivative of the point with respect to the v parameterization coord of the surface

geom: &'a Geometry + 'a

The geometry that was hit

Methods

impl<'a> DifferentialGeometry<'a>

fn new(p: &Point, ng: &Normal, dp_du: &Vector, dp_dv: &Vector, geom: &'a Geometry + 'a) -> DifferentialGeometry<'a>

Setup the differential geometry. Note that the normal will be computed using cross(dp_du, dp_dv)

fn with_normal(p: &Point, n: &Normal, dp_du: &Vector, dp_dv: &Vector, geom: &'a Geometry + 'a) -> DifferentialGeometry<'a>

Setup the differential geometry using the normal passed for the surface normal