Trait tray_rust::geometry::Sampleable

pub trait Sampleable {
    fn sample_uniform(&self, samples: &(f32, f32)) -> (Point, Normal);
    fn sample(&self, p: &Point, samples: &(f32, f32)) -> (Point, Normal);
    fn surface_area(&self) -> f32;
    fn pdf(&self, p: &Point, w_i: &Vector) -> f32;
}

Trait implemented by geometry that can sample a point on its surface

Required Methods

fn sample_uniform(&self, samples: &(f32, f32)) -> (Point, Normal)

Uniformly sample a position and normal on the surface using the samples passed

fn sample(&self, p: &Point, samples: &(f32, f32)) -> (Point, Normal)

Sample the object using the probability density of the solid angle from p to the sampled point on the surface. Returns the sampled point and the surface normal at that point

fn surface_area(&self) -> f32

Return the surface area of the shape

fn pdf(&self, p: &Point, w_i: &Vector) -> f32

Compute the PDF that the ray from p with direction w_i intersects the shape