Module tray_rust::integrator

The integrator module defines the Integrator trait implemented by the various surface integrators used to render the scene with different integration methods, eg. path tracing, photon mapping etc.

Scene Usage Example

The integrator will be specified within the root object of the scene. A type for the integrator along with any additional parameters is required within the integrator object.

"integrator": {
    "type": "The_Integrator_Type",
    ...
}

Reexports

pub use self::whitted::Whitted;

pub use self::path::Path;

pub use self::normals_debug::NormalsDebug;

Modules

normals_debug
Defines the NormalsDebug integrator which renders out the object's normals

path
Defines the Path integrator which implements path tracing with explicit light sampling

whitted
Defines the Whitted integrator which implements Whitted recursive ray tracing See Whitted, An improved illumination model for shaded display

Traits

Integrator
Trait implemented by the various integration methods that can be used to render the scene. For scene usage information see whitted and path to get information on how to specify them.