Struct tray_rust::linalg::normal::Normal

pub struct Normal {
    pub x: f32,
    pub y: f32,
    pub z: f32,
}

Normal is a standard 3 component normal but transforms as a normal normal when transformations are applied

Fields

x

y

z

Methods

impl Normal

fn new(x: f32, y: f32, z: f32) -> Normal

Initialize the normal and set values for x, y, z

fn broadcast(x: f32) -> Normal

Initialize the normal with the same value of x, y, z

fn length_sqr(&self) -> f32

Compute the squared length of the normal

fn length(&self) -> f32

Compute the length of the normal

fn normalized(&self) -> Normal

Get a normalized copy of this normal

fn face_forward(&self, v: &Vector) -> Normal

Return a normal facing along the same direction as v