Struct tray_rust::linalg::transform::Transform

pub struct Transform {
    pub mat: Matrix4,
    pub inv: Matrix4,
}

Transform describes an affine transformation in 3D space and stores both the transformation and its inverse

Fields

mat

inv

Methods

impl Transform

fn identity() -> Transform

Construct the identity transformation

fn from_mat(mat: &Matrix4) -> Transform

Construct a transform from an existing matrix

fn from_pair(mat: &Matrix4, inv: &Matrix4) -> Transform

Construct a transform from an existing matrix/inverse pair

fn translate(v: &Vector) -> Transform

Construct a transformation matrix to translate by the vector

fn scale(v: &Vector) -> Transform

Construct a transform to scale x, y and z by the values in the vector

fn rotate_x(deg: f32) -> Transform

Construct a transform to rotate deg degrees about the x axis

fn rotate_y(deg: f32) -> Transform

Construct a transform to rotate deg degrees about the y axis

fn rotate_z(deg: f32) -> Transform

Construct a transform to rotate deg degrees about the z axis

fn rotate(axis: &Vector, deg: f32) -> Transform

Construct a transform to rotate about axis by deg degrees

fn look_at(pos: &Point, center: &Point, up: &Vector) -> Transform

Construct the look at transform for a camera at pos looking at the point center oriented with up vector up

fn perspective(fovy: f32, near: f32, far: f32) -> Transform

Construct a perspective transformation

fn inverse(&self) -> Transform

Return the inverse of the transformation

fn has_scale(&self) -> bool

Returns true if the transform has applies a scaling

fn inv_mul_point(&self, p: &Point) -> Point

Multiply the point by the inverse transformation TODO: These inverse mults are a bit hacky since Rust doesn't currently have function overloading, clean up when it's added

fn inv_mul_vector(&self, v: &Vector) -> Vector

Multiply the vector with the inverse transformation

fn inv_mul_normal(&self, n: &Normal) -> Normal

Multiply the normal with the inverse transformation

fn inv_mul_ray(&self, ray: &Ray) -> Ray

Multiply the ray with the inverse transformation