Module tray_rust::material::glass

Defines a specular glass material

Scene Usage Example

The specular glass material describes a thin glass surface type of material, not a solid block of glass (there is no absorption of light). The glass requires a reflective and emissive color along with a refrective index, eta.

"materials": [
    {
        "name": "clear_glass",
        "type": "glass",
        "reflect": [1, 1, 1],
        "transmit": [1, 1, 1],
        "eta": 1.52
    },
    ...
]

Structs

Glass
The Glass material describes specularly transmissive and reflective glass material