Trait tray_rust::material::Material

pub trait Material {
    fn bsdf<'a, 'b>(&'a self, hit: &Intersection<'a, 'b>) -> BSDF<'a>;
}

Trait implemented by materials. Provides method to get the BSDF describing the material properties at the intersection

Required Methods

fn bsdf<'a, 'b>(&'a self, hit: &Intersection<'a, 'b>) -> BSDF<'a>

Get the BSDF for the material which defines its properties at the hit point. TODO: When we implement a memory pool we need to pass it here, currently the BxDFs and BSDF are allocated once at surface creation instead of as needed based on material properties.