Struct tray_rust::sampler::adaptive::Adaptive

pub struct Adaptive { /* fields omitted */ }

Adaptive sampler that makes use of the (0, 2) sequence to generate well distributed samples and takes min_spp to max_spp samples per pixel

Methods

impl Adaptive

pub fn new(dim: (u32, u32), min_spp: usize, max_spp: usize) -> Adaptive

Create a low discrepancy sampler to sample the image in dim.0 * dim.1 sized blocks