Struct tray_rust::sampler::block_queue::BlockQueue

pub struct BlockQueue { /* fields omitted */ }

The queue of blocks to be worked on shared immutably between worker threads.

Methods

impl BlockQueue

pub fn new(
img: (u32, u32),
dim: (u32, u32),
select_blocks: (usize, usize)
) -> BlockQueue

Create a block queue for the image with dimensions img. Panics if the image is not evenly broken into blocks of dimension dim

pub fn block_dim(&self) -> (u32, u32)

Get the dimensions of an individual block in the queue

Important traits for BlockQueueIterator<'a>
impl<'a> Iterator for BlockQueueIterator<'a> type Item = (u32, u32);

pub fn iter(&self) -> BlockQueueIterator

Get an iterator to work through the queue

pub fn len(&self) -> usize

Get the length of the queue

pub fn is_empty(&self) -> bool

Check if the queue is empty